requirements.txt
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
tabulate==0.9.0
colorama==0.4.6
ccxt==4.0.3
